refactor(server): separate reconnector code
