+++ apps/server/src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './logger';
+export * from './main-utils';
 export * from './pubsub';
